moved background class
